% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.summary.LRfit}
\alias{print.summary.LRfit}
\title{Print method for summary.LRfit objects}
\usage{
\method{print}{summary.LRfit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{summary.LRfit}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Produces a printed summary of the results for the while-alive loss rate
}
